+++ packages/test-snaps/src/features/snaps/background-events/constants.ts
@@ -1,0 +1,6 @@
+import packageJson from '@metamask/background-events-example-snap/package.json';
+
+export const BACKGROUND_EVENTS_SNAP_ID =
+  'npm:@metamask/background-events-example-snap';
+export const BACKGROUND_EVENTS_SNAP_PORT = 8034;
+export const BACKGROUND_EVENTS_VERSION = packageJson.version;